requirements.txt
django
mysqlclient
djangorestframework
markdown
django-filter
python-decouple
djangorestframework_simplejwt
